+++ src/editor/Groups/rule-types.ts
@@ -3,19 +3,17 @@ import { genValuesWithIncrement } from '../../utils';
 import {
   DEFAULT_ICON_NAME,
   DEFAULT_ICON_RULE_IS_COLLAPSED,
-  DEFAULT_ICON_SIZE,
   DEFAULT_SVG_ICON_V_OFFSET,
 } from 'mapLib/utils';
 
 export interface Rule {
-  lineWidth?: number;
-  nodeSize?: number;
+  width?: number;
+  size?: number;
   overrides?: OverrideTracker | OverField[];
   label: string;
   color?: string;
   collapse?: boolean;
-  iconSize?: number;
-  iconVOffset?: number;
+  offset?: number;
   iconName?: string;
   isEph?: boolean;
   groupIdx?: number;
@@ -25,8 +23,7 @@ export function defaultGroup(label): Rule {
   return {
     label, //: DEFAULT_ICON_RULE_LABEL,
     collapse: DEFAULT_ICON_RULE_IS_COLLAPSED,
-    iconSize: DEFAULT_ICON_SIZE,
-    iconVOffset: DEFAULT_SVG_ICON_V_OFFSET,
+    offset: DEFAULT_SVG_ICON_V_OFFSET,
     iconName: DEFAULT_ICON_NAME,
   };
 }
@@ -50,7 +47,7 @@ export interface OverrideTracker {
   ID: string;
 }
 
-export const NodeSizeStates: SelectableValue[] = genValuesWithIncrement(5, 50, 5, true);
+export const DEFAULT_LINE_WIDTH = 1;
+export const NodeSizeStates: SelectableValue[] = genValuesWithIncrement(5, 50, 5, false);
 export const LineWidthStates: SelectableValue[] = genValuesWithIncrement(0.1, 50, 1, true);
-export const IconSvgSizes: SelectableValue[] = genValuesWithIncrement(10, 150, 5, false);
 export const IconVOffsetValues: SelectableValue[] = genValuesWithIncrement(-20, 20, 5, false);